requirements.txt
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Session==0.3.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
oauthlib==2.1.0
pipenv==2018.11.26
python-dotenv==0.9.1
requests==2.20.1
requests-oauthlib==1.0.0
urllib3==1.24.1
virtualenv==16.1.0
virtualenv-clone==0.4.0
Werkzeug==0.14.1
WTForms==2.2.1
